+++ src/App.tsx
@@ -25,7 +25,7 @@ import TopMenu from 'src/features/TopMenu';
 import VolumeDrawer from 'src/features/Volumes/VolumeDrawer';
 import { requestAccountSettings } from 'src/store/accountSettings/accountSettings.requests';
 import { requestDomains } from 'src/store/domains/domains.actions';
-import { async as imagesAsync } from 'src/store/images';
+import { requestImages } from 'src/store/image/image.requests';
 import { requestLinodes } from 'src/store/linodes/linodes.actions';
 import { requestTypes } from 'src/store/linodeType/linodeType.requests';
 import { requestNotifications } from 'src/store/notification/notification.requests';
@@ -352,7 +352,7 @@ const mapDispatchToProps: MapDispatchToProps<DispatchProps, Props> = (dispatch: 
   return {
     actions: {
       requestDomains: () => dispatch(requestDomains()),
-      requestImages: () => dispatch(imagesAsync.requestImages()),
+      requestImages: () => dispatch(requestImages()),
       requestLinodes: () => dispatch(requestLinodes()),
       requestNotifications: () => dispatch(requestNotifications()),
       requestProfile: () => dispatch(requestProfile()),
